sonarlint refactor p1
