--- src/models/Address.js
@@ -1,12 +1,0 @@
-class Address {
-    constructor(name, companyName, city, district, ward, phoneNo, address){
-        this.name = name;
-        this.companyName = companyName;
-        this.city = city;
-        this.district = district;
-        this.ward = ward; 
-        this.phoneNo = phoneNo; 
-        this.address = address
-    }
-}
-export default Address
