+++ src/store/actions/orderActions.js
@@ -15,11 +15,9 @@ export const getAllOrders = () => async (dispatch) => {
         .get(url)
         .then((res) => {
             dispatch({
-                type: GET_ALL_ORDERS, //this call test dispatch. to dispsatch to our reducer
+                type: GET_ALL_ORDERS,
                 orders: res.data.data,
             });
-
-            // message.success("Got all orders");
         })
         .catch((err) => {
             console.log("Error" + err);
@@ -34,10 +32,9 @@ export const getAllOrdersOfAShop = (shopId) => async (dispatch) => {
         .get(url)
         .then((res) => {
             dispatch({
-                type: GET_ALL_ORDERS_OF_A_SHOP, //this call test dispatch. to dispsatch to our reducer
+                type: GET_ALL_ORDERS_OF_A_SHOP,
                 orders: res.data.data,
             });
-            // message.success("Got orders");
         })
         .catch((err) => {
             console.log("Error" + err);
@@ -52,11 +49,9 @@ export const getOrderById = (orderId) => async (dispatch) => {
         .get(url)
         .then((res) => {
             dispatch({
-                type: GET_SINGLE_ORDER_BY_ID, //this call test dispatch. to dispsatch to our reducer
+                type: GET_SINGLE_ORDER_BY_ID,
                 order: res.data.data,
             });
-
-            // message.success("Got order");
         })
         .catch((err) => {
             console.log("Error" + err);
@@ -71,11 +66,9 @@ export const getOrdersByUserId = (userId) => async (dispatch) => {
         .get(url)
         .then((res) => {
             dispatch({
-                type: GET_MY_ORDERS, //this call test dispatch. to dispsatch to our reducer
+                type: GET_MY_ORDERS,
                 orders: res.data.data,
             });
-
-            // message.success("Got orders");
         })
         .catch((err) => {
             console.log("Error" + err);
@@ -222,7 +215,6 @@ export const deleteOrderById = (orderId) => async () => {
             if (!res.data.success) {
                 return message.error("Error deleting order");
             }
-            // dispatch(getAllOrdersOfAShop(order.shop));
             message.success("Deleted order");
         })
         .catch((err) => {
